requirements.txt
hbutils>=0.14.2
openai
tqdm
pillow
pyyaml
jieba
markdown-it-py
json-repair
pydantic
jinja2
natsort
pathspec
click
pandas